requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
iniconfig==1.1.1
mailjet-rest==1.3.4
Markdown==3.3.3
model-bakery==1.2.1
packaging==20.4
Pillow==8.1.2
pluggy==0.13.1
psycopg2-binary==2.8.6
PyJWT==2.0.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
